Use test directory structure in tracpro.msgs
